backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.0.2
joblib==1.4.2
bcrypt==4.2.0
python-multipart==0.0.12
resend==2.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
PyJWT==2.9.0
